pyproject.toml
[project]
name = "garveyclaw-py"
version = "0.1.0"
description = "Personal AI Agent with Claude Code SDK and Telegram"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.10,<4.0",
    "claude-agent-sdk>=0.1.31",
    "croniter>=6.0.0",
    "lark-oapi>=1.4.0",
    "openai>=2.0.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
]

[project.optional-dependencies]
asr = [
    "vosk>=0.3.45",
]

[project.scripts]
garveyclaw = "garveyclaw.app:main"
garveyclaw-tui = "garveyclaw.tui:main"
garveyclaw-feishu = "garveyclaw.feishu_bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/garveyclaw"]

[tool.ruff]
line-length = 180
